+++ src/main/java/org/folio/innreach/repository/JobExecutionStatusRepository.java
@@ -20,7 +20,7 @@ public interface JobExecutionStatusRepository extends JpaRepository<JobExecution
     "((t.instance_contributed = true and t.updated_date < current_timestamp - (interval '1 hour') * :itemPause) " +
     "or t.instance_contributed = false) and " +
     "t.status in ('READY', 'RETRY') limit :limit) returning * ", nativeQuery = true)
-  List<JobExecutionStatus> updateAndFetchJobExecutionRecordsByStatus(@Param("limit") int limit, @Param("itemPause") int itemPause);
+  List<JobExecutionStatus> updateAndFetchJobExecutionRecordsByStatus(@Param("limit") int limit, @Param("itemPause") double itemPause);
 
   @Query(value = "select count(*) from job_execution_status j inner join contribution c on " +
     "c.job_id = j.job_id where c.status = 0 and j.status='IN_PROGRESS'", nativeQuery = true)
